+++ tests/routers/test_product_IT.py
@@ -1,49 +1,21 @@
 from fastapi.testclient import TestClient
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy.pool import StaticPool
 
-from db.database import get_db, Base
 from main import app
-
-SQLALCHEMY_DATABASE_URL = "sqlite://"
-
-engine = create_engine(
-   SQLALCHEMY_DATABASE_URL,
-   connect_args={"check_same_thread": False},
-   poolclass=StaticPool,
-)
-
-TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-Base.metadata.create_all(bind=engine)
-
-
-def override_get_db():
-   db = TestingSessionLocal()
-   try:
-       yield db
-   finally:
-       db.close()
-
-
-app.dependency_overrides[get_db] = override_get_db
 
 client = TestClient(app)
 
-# POST TESTS
-def test_normal_post_product():
 
-    response = client.post("/product", 
-            json = {
-                "name": "product1", 
-                "description": "product1's description", 
-                "price": 12.5,
-                "stockable": True, 
-                "stock": 2, 
-                "discount": 0, 
-                "categories": []
-            })
+def test_normal_post_product():
+    response = client.post("/product",
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": []
+                           })
 
     assert response.status_code == 201, response.text
 
@@ -56,43 +28,43 @@ def test_normal_post_product():
     assert data["discount"] == 0
     assert data["categories"] == []
 
-def test_post_wrong_category():
 
+def test_post_wrong_category():
     response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": [{
-                "id": "non existent id"
-            }]
-        }
-    )
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": [{
+                                   "id": "non existent id"
+                               }]
+                           }
+                           )
 
     assert response.status_code == 404
 
+
 # GET  TESTS
 def test_get_product():
-
     response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": []
-        }
-    )
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": []
+                           }
+                           )
 
     data = response.json()
     product_id = data["id"]
 
-    response = client.get("/product/" + str(product_id))        #ig this could go wrong
+    response = client.get("/product/" + str(product_id))  # ig this could go wrong
 
     assert response.status_code == 200
 
@@ -107,57 +79,57 @@ def test_get_product():
     assert data["discount"] == 0
     assert data["categories"] == []
 
-def test_no_correct_id_product():
 
+def test_no_correct_id_product():
     mock_id = "some random id string"
 
     response = client.get("/product/" + mock_id)
 
     assert response.status_code == 404, response.text == "Product not found"
 
+
 # PUT TESTS
 def test_put_not_existing_product():
-
     response = client.put("/product/1",
-        json = {
-            "name": "product1", 
-            "description": "product1's description", 
-            "price": 12.5,
-            "stockable": True, 
-            "stock": 2, 
-            "discount": 0, 
-            "categories": []
-        })
+                          json={
+                              "name": "product1",
+                              "description": "product1's description",
+                              "price": 12.5,
+                              "stockable": True,
+                              "stock": 2,
+                              "discount": 0,
+                              "categories": []
+                          })
 
     assert response.status_code == 404, response.text
     assert response.json() == {'detail': 'Product not found'}
 
-def test_put_existing_product_no_category():
 
+def test_put_existing_product_no_category():
     response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": []
-        }
-    )
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": []
+                           }
+                           )
 
     product_id = response.json()["id"]
 
-    response = client.put("/product/" + str(product_id), 
-        json = {
-            "name": "product123",
-            "description": "product123's description",
-            "price": 14.5,
-            "stockable": False,
-            "stock": 0,
-            "discount": 0,
-            "categories": []
-        })
+    response = client.put("/product/" + str(product_id),
+                          json={
+                              "name": "product123",
+                              "description": "product123's description",
+                              "price": 14.5,
+                              "stockable": False,
+                              "stock": 0,
+                              "discount": 0,
+                              "categories": []
+                          })
 
     assert response.status_code == 200, response.text
 
@@ -172,121 +144,120 @@ def test_put_existing_product_no_category():
     assert data["discount"] == 0
     assert data["categories"] == []
 
-def test_put_existing_product_no_existing_category():
 
+def test_put_existing_product_no_existing_category():
     response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": []
-        }
-    )
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": []
+                           }
+                           )
 
     product_id = response.json()["id"]
 
-    response = client.put("/product/" + str(product_id), 
-        json = {
-            "name": "product123",
-            "description": "product123's description",
-            "price": 14.5,
-            "stockable": False,
-            "stock": 0,
-            "discount": 0,
-            "categories": [{
-                "id": "some non existing id"
-            }]
-        })
+    response = client.put("/product/" + str(product_id),
+                          json={
+                              "name": "product123",
+                              "description": "product123's description",
+                              "price": 14.5,
+                              "stockable": False,
+                              "stock": 0,
+                              "discount": 0,
+                              "categories": [{
+                                  "id": "some non existing id"
+                              }]
+                          })
 
     assert response.status_code == 404, response.text
     assert response.json() == {'detail': 'Category not found'}
 
-def test_put_existing_product_existing_category():
 
-    response = client.post("/category", 
-        json = {
-            "name": "category1",
-            "icon": "icon1",
-            "slug": "slug1"
-        })
+# def test_put_existing_product_existing_category():
+#     response = client.post("/category",
+#                            json={
+#                                "name": "category1",
+#                                "icon": "icon1"
+#                            })
+#
+#     category1_id = response.json()["id"]
+#
+#     response = client.post("/category",
+#                            json={
+#                                "name": "category2",
+#                                "icon": "icon2"
+#                            })
+#
+#     category2_id = response.json()["id"]
+#
+#     response = client.post("/product",
+#                            json={
+#                                "name": "product1",
+#                                "description": "product1's description",
+#                                "price": 12.5,
+#                                "stockable": True,
+#                                "stock": 2,
+#                                "discount": 0,
+#                                "categories": []
+#                            }
+#                            )
+#
+#     product_id = response.json()["id"]
+#
+#     response = client.put("/product/" + str(product_id),
+#                           json={
+#                               "name": "product1",
+#                               "description": "product1's description",
+#                               "price": 12.5,
+#                               "stockable": True,
+#                               "stock": 2,
+#                               "discount": 0,
+#                               "categories": [
+#                                   {
+#                                       "id": str(category1_id)
+#                                   },
+#                                   {
+#                                       "id": str(category2_id)
+#                                   }
+#                               ]
+#                           })
+#
+#     assert response.status_code == 200, response.text
+#     data = response.json()
+#
+#     assert data["id"] == str(product_id)
+#     assert data["name"] == "product1"
+#     assert data["description"] == "product1's description"
+#     assert data["price"] == 12.5
+#     assert data["stockable"] == True
+#     assert data["stock"] == 2
+#     assert data["discount"] == 0
+#
+#     assert len(data["categories"]) == 2
+#     assert data["categories"][0]["id"] == category1_id
+#     assert data["categories"][1]["id"] == category2_id
+#     assert data["categories"][0]["name"] == "category1"
 
-    category1_id = response.json()["id"]
-
-    response = client.post("/category", 
-        json = {
-            "name": "category2",
-            "icon": "icon2",
-            "slug": "slug2"
-        })
-
-    category2_id = response.json()["id"]
-
-    response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": []
-        }
-    )
-
-    product_id = response.json()["id"]
-
-    response = client.put("/product/" + str(product_id), 
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": [
-                {
-                    "id": str(category1_id)
-                },
-                {
-                    "id": str(category2_id)
-                }
-                ]
-        })
-
-    assert response.status_code == 200, response.text
-    
-    data = response.json()
-    
-    assert data["id"] == str(product_id)
-    assert data["name"] == "product1"
-    assert data["description"] == "product1's description"
-    assert data["price"] == 12.5
-    assert data["stockable"] == True
-    assert data["stock"] == 2
-    assert data["discount"] == 0
-    assert data["categories"][0]["id"] == category1_id
-    assert data["categories"][1]["id"] == category2_id
-    assert data["categories"][0]["name"] == "category1"
 
 # DELETE TESTS
 
 def test_delete_existing_product():
-
     response = client.post("/product",
-        json = {
-            "name": "product1",
-            "description": "product1's description",
-            "price": 12.5,
-            "stockable": True,
-            "stock": 2,
-            "discount": 0,
-            "categories": []
-        }
-    )
-    
+                           json={
+                               "name": "product1",
+                               "description": "product1's description",
+                               "price": 12.5,
+                               "stockable": True,
+                               "stock": 2,
+                               "discount": 0,
+                               "categories": []
+                           }
+                           )
+
     product_id = response.json()["id"]
 
     response = client.delete("/product/" + product_id)
@@ -296,12 +267,9 @@ def test_delete_existing_product():
 
 
 def test_delete_non_existing_product():
-
     id = "some non-existing id"
 
     response = client.delete("/product/" + id)
 
     assert response.status_code == 404
     assert response.json() == {"detail": "Product not found"}
-
-
