+++ tests/test_sql_app.py
@@ -1,0 +1,29 @@
+from sqlalchemy import create_engine, StaticPool
+from sqlalchemy.orm import sessionmaker
+from starlette.testclient import TestClient
+
+from db.database import Base, get_db
+from main import app
+
+SQLALCHEMY_DATABASE_URL = "sqlite://"
+
+engine = create_engine(
+    SQLALCHEMY_DATABASE_URL,
+    connect_args={"check_same_thread": False},
+    poolclass=StaticPool,
+)
+
+TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+Base.metadata.create_all(bind=engine)
+
+
+def override_get_db():
+    db = TestingSessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+app.dependency_overrides[get_db] = override_get_db
